Make `list` generator much faster
We now use a single bit for the booleans. This changes the behaviour
somewhat, but not in an important manner.

`tasty` now no longer feels the need to tell us how slow the `list` test
is; it started at 0.4 seconds before this work, down to 0.03 seconds
after we used smaller precision for the list elements, and now no time
at all once now that we use 1 bit for the boolean.

Closes #3.