Improve various aspects of the design
As we are getting more experience with this approach, some patterns are
starting to emerge. One important lesson is: do not generate values that
you are not using (in the hope of using them later as some other value
shrinks): the unused value can then always trivially shrink towards
zero before it is ever used.

This also means that we cannot for example generate a complete tree and
then pick a subtree from it: generation will be linear in the size of
the subtree, but shrinking will /not/ be! This is demonstrated in
`TestSuite.Sanity.Tree`.