Introduce `TestRun`
This continues the "How To Specify It" paper, but it has one check that
says

```haskell
prop_valid_nil :: Property ()
prop_valid_nil = assert "" $ valid (nil :: BST Int Int)
```

This is a property that doesn't generate anything and so there is no
point running 100 tests. This situation is now detected, and only one
test is run.

This also moves most of the `Property` stuff to an internal module,
exposing only the public API in `Test.Falsify.Property`.