Generate and shrink functions
This also fixes the shrinker for bool, which was using the full 64-bit
precision, where 1 sufficed. This significantly improves the performance of the
generator.

Performance is not great; but I think this is a general problem with
`fromShrinkTree`. Perhaps we can improve this by changing shrinking so that
along with a new sample tree, we also get a new generator. This might also give
us a nicer way to deal with initializing the sample tree.

Closes #11.