Significantly improve performance of `bst`
This an interesting observation: instead of producing a `Tree a` directly, we
effectively first generate a `Tree (Gen a)`, and then execut the effects as a
separate step. This changes the order in which shrinking happens: rather than
trying to shrink individual elements first, we try to shrink entire subtrees
first.