Simply the definition of `Gen`
This moves away from the free definition; I didn't see how do this previously;
the correct type of `Gen` eluded me. In hindsight, it's kind of obvious
however; it feels quite natural now. The new definition is improves separation
of concerns (for example, `Selective` is just a single new instance; there is
no need to modify both a `gen` and a `shrink` function), and at the same time
more expressive (a direct result of this now being a deeper embedding).

This also makes it possible to discard tests; at first I thought we could do
this by changing `Gen` to have the type

```haskell
newtype Gen a = Gen (SampleTree -> Maybe (a, Truncated, [SampleTree]))
```

with a `Maybe` in the result. However, this is incompatible with the generation
of infinite data structures.