Fix testMinimum
If we are testing the minimum value, we should not look for arbitrary paths
through the shrink tree, but rather the optimum one. In other words, we should
be doing precisely what `shrinkFrom` does! So `testMinimum` is now defined in
terms of the real shrinker, which interestingly means that `testMinimum` can
replace some of our sanity checks, and replace them with property tests. Nice.

The shuffle property is still failing, but I think it's now pretty clear that
the remaining problem is related to shuffle, not the infrastructure. The
property is too strong.