Remove support for `Truncated`
This was useful in the initial design stages of the library, when we were still
trying to understand how things were working. However, it was difficult to
maintain, not all combinators were supporting it properly anyway, and the tests
that were using it can now be replaced by more powerful property tests (the
library testing itself).